fixed the out of bould bug

+++ src/rendering/rendering.c
@@ -6,7 +6,7 @@
 /*   By: mlendle <mlendle@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/23 14:40:09 by mlendle           #+#    #+#             */
-/*   Updated: 2025/07/03 12:02:20 by mlendle          ###   ########.fr       */
+/*   Updated: 2025/07/03 12:10:14 by mlendle          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static uint32_t	get_wall_pixel_color(t_game *game, t_ray *ray, int pixel_y)
 {
 	mlx_texture_t	*texture;
 	double			wall_hit_pos;
-	double			height;
+	int				height;
 	int				tex_x;
 	int				tex_y;
 
@@ -46,17 +46,8 @@ static uint32_t	get_wall_pixel_color(t_game *game, t_ray *ray, int pixel_y)
 	if (ray->hit_dir == WALL_WEST || ray->hit_dir == WALL_SOUTH)
 		wall_hit_pos = 1.0 - wall_hit_pos;
 	tex_x = (int)(wall_hit_pos * texture->width);
-	// if (tex_x < 0)
-	// 	tex_x = 0;
-	// else if (tex_x >= (int)texture->width)
-	// 	tex_x = texture->width - 1;
 	tex_y = (int)(((double)(pixel_y - (game->mlx->height - height) / 2)
 				/ height) * texture->height);
-	// we shouldn't get any out of bounds values
-	// if (tex_y < 0)
-	// 	tex_y = 0;
-	// else if (tex_y >= (int)texture->height)
-	// 	tex_y = texture->height - 1;
 	return (get_pixel_color(texture, tex_x, tex_y));
 }
